The Log_file class has been completely rewritten based on feedback from Roman Neuhauser.
Log output is no longer buffered before being written to the log file; it
is now written immediately.  The time format, the end-on-line sequence and
the overall format of the log line are now configurable.  It is also possible
to specify whether or not new log lines will be append to an existing log
file (as opposed to overwriting the old file with a new one).